change zrange method min, max params from int to double

+++ qtredis.h
@@ -120,7 +120,7 @@ public:
     QStringList sort(QString key, QString conditions);
     int ttl(QString key);
     QString type(QString key);
-    QStringList zrange(QString key, int max, int min, QString other="");
+    QStringList zrange(QString key, double max, double min, QString other="");
 
 signals:
     void connected();
